Rename exceptions package
