+++ marketplace-server/src/main/java/com/ncgroup/marketplaceserver/goods/service/GoodsService.java
@@ -3,7 +3,7 @@ package com.ncgroup.marketplaceserver.goods.service;
 import com.ncgroup.marketplaceserver.goods.exceptions.GoodAlreadyExistsException;
 import com.ncgroup.marketplaceserver.goods.model.Good;
 import com.ncgroup.marketplaceserver.goods.model.GoodDto;
-import com.ncgroup.marketplaceserver.shopping.cart.exceptions.NotFoundException;
+import com.ncgroup.marketplaceserver.shopping.cart.exception.NotFoundException;
 
 import java.util.*;
