+++ marketplace-server/src/main/java/com/ncgroup/marketplaceserver/shopping/cart/exception/AccessDeniedException.java
@@ -1,4 +1,4 @@
-package com.ncgroup.marketplaceserver.shopping.cart.exceptions;
+package com.ncgroup.marketplaceserver.shopping.cart.exception;
 
 public class AccessDeniedException extends Exception{
     public AccessDeniedException(String message) {
