+++ marketplace-server/src/main/java/com/ncgroup/marketplaceserver/shopping/cart/exception/NotFoundException.java
@@ -1,4 +1,4 @@
-package com.ncgroup.marketplaceserver.shopping.cart.exceptions;
+package com.ncgroup.marketplaceserver.shopping.cart.exception;
 
 public class NotFoundException extends Exception{
     public NotFoundException(String message) {
