+++ marketplace-server/src/main/java/com/ncgroup/marketplaceserver/shopping/cart/service/ShoppingCartItemService.java
@@ -1,14 +1,10 @@
 package com.ncgroup.marketplaceserver.shopping.cart.service;
 
-import com.ncgroup.marketplaceserver.security.model.UserPrincipal;
-import com.ncgroup.marketplaceserver.shopping.cart.exceptions.AccessDeniedException;
-import com.ncgroup.marketplaceserver.shopping.cart.exceptions.NotFoundException;
-import com.ncgroup.marketplaceserver.shopping.cart.model.ShoppingCartItem;
+import com.ncgroup.marketplaceserver.shopping.cart.exception.NotFoundException;
 import com.ncgroup.marketplaceserver.shopping.cart.model.dto.ShoppingCartItemCreateDto;
 import com.ncgroup.marketplaceserver.shopping.cart.model.dto.ShoppingCartItemReadDto;
 import com.ncgroup.marketplaceserver.shopping.cart.model.dto.ShoppingCartItemUpdateDto;
 
-import java.util.Collection;
 import java.util.List;
 
 public interface ShoppingCartItemService {
