+++ marketplace-server/src/main/java/com/ncgroup/marketplaceserver/shopping/cart/service/ShoppingCartItemServiceImpl.java
@@ -2,7 +2,7 @@ package com.ncgroup.marketplaceserver.shopping.cart.service;
 
 import com.ncgroup.marketplaceserver.model.User;
 import com.ncgroup.marketplaceserver.service.UserService;
-import com.ncgroup.marketplaceserver.shopping.cart.exceptions.NotFoundException;
+import com.ncgroup.marketplaceserver.shopping.cart.exception.NotFoundException;
 import com.ncgroup.marketplaceserver.shopping.cart.model.ShoppingCartItem;
 import com.ncgroup.marketplaceserver.shopping.cart.model.dto.ShoppingCartItemCreateDto;
 import com.ncgroup.marketplaceserver.shopping.cart.model.dto.ShoppingCartItemReadDto;
@@ -13,7 +13,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
